requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
gensim==4.1.2
idna==3.3
joblib==1.1.0
nltk==3.6.5
numpy==1.21.4
pandas==1.3.4
pyaml==21.10.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.1
scikit-optimize==0.9.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
